Refactoed Trainer.py + Added Trainer sub classes
The idea is that the base Trainer class stores the data and such, the sub-classes define how each round of training works. The base class also takes care of normalization, loss tracking (or, really just setting up the stuff for this), checkpointing, and timing. The sub class specifies how each round of training happens/defines the optimizer/defines the loss functions and such.
